Break code classes to their own files
