+++ .travis.yml
@@ -7,16 +7,13 @@ matrix:
     sudo: required
     services:
       - mysql
+    before_install:
+      - sudo mysql -e "use mysql; update user set authentication_string=PASSWORD('') where User='root'; update user set plugin='mysql_native_password';FLUSH PRIVILEGES;"
     before_script:
       - curl -s http://getcomposer.org/installer | php
       - php composer.phar install
       - mysql -V
       - mysqldump -V
-      - tests/create_users.sh
+      - tests/scripts/create_users.sh
     script:
-      - php -l src/Ifsnop/Mysqldump/Mysqldump.php
-      - php src/Ifsnop/Mysqldump/Mysqldump.php
-      - vendor/bin/phpunit
       - cd tests && ./test.sh
-    before_install:
-      - sudo mysql -e "use mysql; update user set authentication_string=PASSWORD('') where User='root'; update user set plugin='mysql_native_password';FLUSH PRIVILEGES;"
