+++ src/Ifsnop/Mysqldump/Compress/CompressInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Ifsnop\Mysqldump\Compress;
+
+interface CompressInterface
+{
+    public function open(string $filename);
+
+    public function write(string $str): int;
+
+    public function close(): bool;
+}
