+++ src/Ifsnop/Mysqldump/TypeAdapter/TypeAdapterDblib.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Ifsnop\Mysqldump\TypeAdapter;
+
+class TypeAdapterDblib extends TypeAdapterFactory
+{
+}
