+++ src/Ifsnop/Mysqldump/TypeAdapter/TypeAdapterPgsql.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Ifsnop\Mysqldump\TypeAdapter;
+
+class TypeAdapterPgsql extends TypeAdapterFactory
+{
+}
