+++ src/Ifsnop/Mysqldump/TypeAdapter/TypeAdapterSqlite.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Ifsnop\Mysqldump\TypeAdapter;
+
+class TypeAdapterSqlite extends TypeAdapterFactory
+{
+}
